Update the rest of the Doubles to Ints.

+++ WalletLibrary/Submodules/VerifiableCredential-SDK-iOS/VCEntities/VCEntities/linkedDomains/DomainLinkageCredential.swift
@@ -10,7 +10,7 @@
 struct DomainLinkageCredentialClaims: Claims {
     let subject: String
     let issuer: String
-    let notValidBefore: Double
+    let notValidBefore: Int
     let verifiableCredential: DomainLinkageCredentialContent
     
     enum CodingKeys: String, CodingKey {
